collected similar and duplcated code in `writers.py`
