+++ src/napari_omero/plugins/omero.py
@@ -12,19 +12,13 @@ import omero.clients
 from napari_omero.utils import lookup_obj, obj_to_proxy_string
 from omero.cli import CLI, BaseControl, ProxyStringType
 from omero.gateway import BlitzGateway, PixelsWrapper
-from omero.model import (
-    EllipseI,
-    ImageI,
-    LineI,
-    PointI,
-    PolygonI,
-    PolylineI,
-    RectangleI,
-    RoiI,
-)
-from omero.rtypes import rdouble, rint, rstring
 
-from .masks import save_labels
+from .writers import (
+    save_labels,
+    create_omero_point,
+    create_omero_shape,
+    create_roi
+    )
 
 HELP = "Connect OMERO to the napari image viewer"
 
@@ -160,7 +154,11 @@ def set_dims_defaults(viewer, image):
         viewer.dims.set_point(1, image.getDefaultZ())
 
 
-def save_rois(viewer, image):
+def save_rois(
+        viewer,
+        image,
+        mode: str = "Update"
+    ) -> None:
     """Save napari ROIs to OMERO.
 
     Usage: In napari, open console...
@@ -177,7 +175,8 @@ def save_rois(viewer, image):
     roi_service = conn.getRoiService()
     roi_ids = roi_service.findByImage(image.getId(), None).rois
 
-    # Loop through all ROIs
+    # Loop through all ROIs:
+    # We need to know what ROIs/shapes already exist on the remote
     remote_roi_ids = []
     remote_shape_ids = []
 
@@ -199,7 +198,7 @@ def save_rois(viewer, image):
                 shape_id_local = layer.features.iloc[idx]["shape_id"]
                 if pd.isna(shape_id_local) and pd.isna(roi_id_local):
                     point = create_omero_point(p)
-                    roi = create_roi(conn, image.id, [point])
+                    roi = create_roi(image, [point])
                     print(f"Created ROI: {roi.id.val}")
 
             # remove any remote shapes that are not present locally
@@ -223,7 +222,7 @@ def save_rois(viewer, image):
                 if pd.isna(shape_id_local) and pd.isna(roi_id_local):
                     shape = create_omero_shape(shape_type, data)
                     if shape is not None:
-                        roi = create_roi(conn, image.id, [shape])
+                        roi = create_roi(image, [shape])
                         print(f"Created ROI: {roi.id.val}")
 
             # remove any remote shapes that are not present locally
@@ -235,100 +234,6 @@ def save_rois(viewer, image):
         elif type(layer) is labels_layer:
             print("Saving Labels...")
             save_labels(layer, image)
-
-
-def get_x(coordinate):
-    return coordinate[-1]
-
-
-def get_y(coordinate):
-    return coordinate[-2]
-
-
-def get_t(coordinate):
-    return coordinate[0]
-
-
-def get_z(coordinate):
-    return coordinate[1]
-
-
-def create_omero_point(data):
-    point = PointI()
-    point.x = rdouble(get_x(data))
-    point.y = rdouble(get_y(data))
-    point.theZ = rint(get_z(data))
-    point.theT = rint(get_t(data))
-    return point
-
-
-def create_omero_shape(shape_type, data):
-    # "line", "path", "polygon", "rectangle", "ellipse"
-    # NB: assume all points on same plane.
-    # Use first point to get Z and T index
-    z_index = get_z(data[0])
-    t_index = get_t(data[0])
-    shape = None
-    if shape_type == "line":
-        shape = LineI()
-        shape.x1 = rdouble(get_x(data[0]))
-        shape.y1 = rdouble(get_y(data[0]))
-        shape.x2 = rdouble(get_x(data[1]))
-        shape.y2 = rdouble(get_y(data[1]))
-    elif shape_type in ["path", "polygon"]:
-        shape = PolylineI() if shape_type == "path" else PolygonI()
-        # points = "10,20, 50,150, 200,200, 250,75"
-        points = [f"{get_x(d)},{get_y(d)}" for d in data]
-        shape.points = rstring(", ".join(points))
-    elif shape_type in ["rectangle", "ellipse"]:
-        # corners go anti-clockwise starting top-left
-        x1 = get_x(data[0])
-        x2 = get_x(data[1])
-        x3 = get_x(data[2])
-        x4 = get_x(data[3])
-        y1 = get_y(data[0])
-        y2 = get_y(data[1])
-        y3 = get_y(data[2])
-        y4 = get_y(data[3])
-        if shape_type == "rectangle":
-            # Rectangle not rotated
-            if x1 == x2:
-                shape = RectangleI()
-                # TODO: handle 'updside down' rectangle x3 < x1
-                shape.x = rdouble(x1)
-                shape.y = rdouble(y1)
-                shape.width = rdouble(x3 - x1)
-                shape.height = rdouble(y2 - y1)
-            else:
-                # Rotated Rectangle - save as Polygon
-                shape = PolygonI()
-                points_str = f"{x1},{y1}, {x2},{y2}, {x3},{y3}, {x4},{y4}"
-                shape.points = rstring(points_str)
-        elif shape_type == "ellipse":
-            # Ellipse not rotated (ignore floating point rouding)
-            if int(x1) == int(x2):
-                shape = EllipseI()
-                shape.x = rdouble((x1 + x3) / 2)
-                shape.y = rdouble((y1 + y2) / 2)
-                shape.radiusX = rdouble(abs(x3 - x1) / 2)
-                shape.radiusY = rdouble(abs(y2 - y1) / 2)
-            else:
-                # TODO: Need to calculate transformation matrix
-                print("Rotated Ellipse not yet supported!")
-
-    if shape is not None:
-        shape.theZ = rint(z_index)
-        shape.theT = rint(t_index)
-    return shape
-
-
-def create_roi(conn, img_id, shapes):
-    updateService = conn.getUpdateService()
-    roi = RoiI()
-    roi.setImage(ImageI(img_id, False))
-    for shape in shapes:
-        roi.addShape(shape)
-    return updateService.saveAndReturnObject(roi, conn.SERVICE_OPTS)
 
 
 class NonCachedPixelsWrapper(PixelsWrapper):
